Update Items
- changed int to Long for primary key
- made changes to implement description

+++ src/main/java/com/csit321g2/Capstone/Repository/ItemsRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.csit321g2.Capstone.Entity.ItemsEntity;
 
 @Repository
-public interface ItemsRepository extends JpaRepository<ItemsEntity, Integer>{
+public interface ItemsRepository extends JpaRepository<ItemsEntity, Long>{
 	
 }
